Add a job to expire customer's order in 20 minutes
This is done via Active Job API. The job will be queued to run in 20
minutes after user modified the order (adding item to cart, going
through checkout flow, etc).

If the user modifies the after the job is queued, the job will become
stale and it will do nothing with the order once the job executes. Also,
if the order completes, the job will become stale as well.